+++ laravel/database/seeders/DatabaseSeeder.php
@@ -15,6 +15,7 @@ class DatabaseSeeder extends Seeder
     public function run(): void
     {
         $this->call([
+            BootstrapUserSeeder::class,
             CurrencySeeder::class,
             RbacSeeder::class,
         ]);